+++ src/Factory/PointFactory.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Location\Factory;
+
+class PointFactory extends CoordinateFactory
+{
+}