+++ src/Point.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Location;
+
+class Point extends Coordinate
+{
+}